refactor: split SettingsSubscreens.swift into one file per sub-screen (#80)
SettingsSubscreens.swift held nine independent sub-screen Views (1242 lines)
with no shared state — a Single-Responsibility violation flagged by #80's
file_length list. Decomposed along the natural seam (one screen per file):

  NameEditView, HeightPickerView, BirthdayPickerView, GenderPickerView,
  ThemePickerView, CadreSyncView, ExportCSVView, ImportCSVView, AboutCadreView

Two file-private helpers move with their sole consumer: ExportItem →
ExportCSVView, the CSVFormat.displayName extension → ImportCSVView.

Pure code move — no behavior change. Each View is verbatim; only the
per-file imports were narrowed to what each screen actually uses
(SwiftData for the two CSV screens, UniformTypeIdentifiers for import).

Verified: make lint (0 violations, --strict), make build, make test
(Baseline-CI plan) all green, including testSettingsScreenAccessibility,
CSVImportFlowUITests, and IdentifierCoverageUITests.

Note: file_length is still dominated by TrendsView.swift (1648), so the
.swiftlint.yml file_length threshold can't ratchet down from this change
alone — that lands with the TrendsView decomposition.
